flashchips.c: import generic/unknown Winbond chip from upstream
This chip was previously imported in
`commit 94a80d615f2e1c88a763202d091732e2bdfa3a03` along with a generic
SFDP-capable chip. The SFDP chip caused a regression where multiple
chips were identified and the patch was reverted as an immediate fix.
The issue with the SFDP chip was resolved later, so this patch just
imports the generic Winbond chip definition again.

To ensure this doesn't cause the same type of regression that we saw
with the SFDP chip, I tested flashrom with a W25X40 to verify that the
generic chip definition isn't used if the chip ID matches a specific
flashchip instead.

BUG=b:166294558
TEST=builds
BRANCH=none

Change-Id: I31a6499a86f737b2c45a4cb749ae41a71ca525d0
Signed-off-by: Nikolai Artemiev <nartemiev@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2807219
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>