writeprotect.c: delete {get,set}_modifer_bits() from wp_context
Delete the {get,set}_modifer_bits function pointers in struct
wp_context. They were used to provide access to the TB/SEC bits in a
chip-specific way. This may appear to have been a good abstraction, but
it was only used by a small set of S25F chips, to effectively override
the normal code paths in writeprotect.c for accessing the SEC/TB bits.

This patch rewrites the calls to s25f_{get,set}_modifer_bits() as the
hard-coded special cases that they actually are. This will allow
wp_context to be deleted and a proper abstraction to be found in the
future.

BUG=b:182223106
TEST=builds
BRANCH=none

Signed-off-by: Nikolai Artemiev <nartemiev@google.com>
Change-Id: I378e610dbf1040f2144528c920d979332e6267c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2816708
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>