meson: Define USE_BIG_LOCK=1
Prior to Makefile->meson transition, we used to set USE_BIG_LOCK=1.

This accidentally fixes an issue when flashrom is called with
stdout (fd=1) closed (which cgpt does -- that's wrong and we'll fix it).
With USE_BIG_LOCK=1 the lockfile takes up fd=1, and console output is
(incorrectly) sent to it. But that's better than corrupting the flash.

Also, AFAICT, it was not intentional to drop USE_BIG_LOCK=1, so let's
re-instate it, at least for the time being.

BUG=b:184559695
TEST=`bash -c "cd /tmp; exec 1>&-; /usr/sbin/flashrom -i RW_GPT:rw_gpt -r"`
     check that flash is not corrupted
     `flashrom -r /usr/local/x.bin && hexdump -C /usr/local/x.bin | head`
BRANCH=none

Change-Id: Ia932f5750f808b213a46582943d77ac313a220ec
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Exempt-From-Owner-Approval: Breaks devices in the lab, see bug.
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2812287
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Daniel Campello <campello@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>