flash.h: Change do_diff flag to do_not_diff
This changes makes the default flag state to do_not_diff=0.

Upstream does not have this flag as it is action_descriptor dependent,
so libflashrom does not have an interface to set/unset the flag.
Without this change libflashrom defaults to do_diff=0 and fail to
write/erase if the chip is not empty.

BRANCH=none
BUG=b:180556238,b:184740023
TEST=flashrom_image_write()

Change-Id: I6b206474a2716b701f5462619d906b027a1655bc
Signed-off-by: Daniel Campello <campello@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2810151
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>