flashrom.c: modernize do_extract() to use do_read()
This change modernizes do_extract() so it uses the do_read() path which
includes read_from_layout() and enables for a partial read of the flash
contents according to the given layout. It also fixes the
prepare_flash_access flags used (as a read operation is performed) and
it respects the filenames that were provided via -i region:filename.

BRANCH=none
BUG=b:185614763
TEST=flashrom -p lspcon_i2c_spi:bus=7 -l layout -i FLAG:f.txt -x

Change-Id: I2ec440a6600d236f46bc401421167a7296971278
Signed-off-by: Daniel Campello <campello@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2832013
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>