cli_classic.c: add parse_wp_range()
Current behavior is implemented by looking at the remaning arguments
after parsing all CLI options. For wp-range there is an expectation of 2
additional arguments. With this change we allow for only a single
argument in addition of the two. Conditional on the amount of extra
arguments provided we parse the range (i.e., for one additional argument
we parse as <start>,<len> and for two we do as before <start> <len>)

This backports d3b6acffe40351d84cfd6aaa0238f9a653d69aa1

BRANCH=none
BUG=b:186696898
TEST=flashrom -p host --wp-range 0x0,0x0; flashrom -p host --wp-range 0x0 0x0

Change-Id: Id8b2b243eb16f8f4666107d5256d906df6590553
Signed-off-by: Daniel Campello <campello@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2857373
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>