flashrom.c: only align reads when building descriptor
With this change we avoid reading the extra contents from aligning to
block erase boundaries to the buffer provided by do_read() and
flashrom_image_read().

The way that read_by_layout() is currently implemented downstream is
that it pads the read with enough data to complete an erase block.
This is used when writing, so after erasing the block we can mix the
new contents with existent contents to fill the block that is erased.

This is not needed when just reading and it is actually bad when
reading a small region via -i (instead of the whole contents).
As an example, if you use do_read or flashrom_image_read for a 1-byte
size region, then the current implementation will fill up the provided
buffer with a whole erase block worth of data instead of the 1-byte
requested.

BUG=b:180556238,b:186215011
BRANCH=none
TEST=emerge flashrom; flashrom -r

Signed-off-by: Daniel Campello <campello@chromium.org>
Change-Id: I873e2b9bbccc88bc9a1983684a980cb168eb9053
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2841739
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>