flashrom: remove opaque chip status register indirection
This patch removes several layers of indirection that were used by
ichspi to provide access to the status register. Access is now provided
by ichspi directly populating fields in the opaque chip during probing.
This is consistent with other uses of the opaque chip, though still
capable of being improved.

This commit allows opaque.c, flashchips.c, and the definition of
`struct opaque_master` in programmer.h to be resynced with upstream.

BUG=b:166294558
BRANCH=none
TEST=tested on bobba360, added logging to {read,write}_status_opaque()
     to check that they were still being called as expected.

Signed-off-by: Nikolai Artemiev <nartemiev@google.com>
Change-Id: I5dd517a1949bbbed186df54cc5c6b64a1496b3c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/2882241
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>