UPSTREAM: ft2232_spi: Add FTDI search by description.
This adds to the search-by-serial functionality with
search-by-description (product string). This is useful when e.g. one has
multiple FTDIs in a system and wants the serial numbers to reflect the
system-level serial number, and the description to reflect the
subcomponent names.

Tested manually by running with both serial and description searches, on
a machine with multiple FTDIs plugged in. Ensured that when two devices
with the same vid/pid/serial number are plugged in, description can
be used to differentiate.

Verifed no-description, no-serial, one FTDI plugged in base case works.

Original version of this code used the original single "arg" char*, but
on further thought, this wasn't worth the readability and functionality
losses. The new version with arg2 gets rid of several lines of code, the
gotos, and adds the ability to filter by both description and serial
simultaneously.

BUG=none
BRANCH=none
TEST=none

Original-Change-Id: Ib4be23247995710900175f5f16e38db577ef08fa
Original-Signed-off-by: Harry Johnson <johnsonh@waymo.com>
Original-Reviewed-on: https://review.coreboot.org/c/flashrom/+/56164
Original-Reviewed-by: Anastasia Klimchuk <aklm@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
(cherry picked from commit b7dec7f2cce9f81d5059c3c3e0d65d14fd9363c8)

Change-Id: I66c4d4cb69407229ba93eb1cb66cf4a57ccb7ac1
Signed-off-by: Anastasia Klimchuk <aklm@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/3113388
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>