wp_chip_info: add per-chip writeprotect register info
Add a bitmask for each chip indicating which bits in SR1 have an effect
on the selected memory protection range.

In addition to the bits in SR1, there are a few bits in other registers
that affect the selected range. There are only 3 such bits, so they are
just represented as special-case flags for each chip.

BUG=b:182223106
BRANCH=none
TEST=builds

Signed-off-by: Nikolai Artemiev <nartemiev@google.com>
Change-Id: Ib69bc29232e9aa6669eabb806a361dbe8ffa1f3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/3017171
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>