writeprotect: Fix MX25U12835E
Due to various bits of writeprotect code being reused with incompatible
chips, the range tables for the MX25U12835E chip were misinterpreted,
making some ranges unavailable.

The intended behavior was that the TB bit would be read from the MXCR
register and then used to select a table of ranges that are not filtered
any further. However due to reusing wp code for other chips, an
unrelated SR1 bit is also interpreted as the TB bit incorrectly filters
out valid ranges. Fix it for now with another hack: set the TB bit in
the tables to "DON'T CARE".

BUG=b:182223106
BRANCH=none
TEST=ninja test

Signed-off-by: Nikolai Artemiev <nartemiev@google.com>
Change-Id: I2e4a3cb5c5020cecc3546f3fe3bb62f026c3bf64
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/3067493
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>