UPSTREAM: ich_descriptors.c Invert the meaning of 'dual_output' bit
In the Flash Component description register (FLCOMP) bit 30 reports the
capability of using dual output for fast read operation on the flash
component. According to various SPI Programming Guides (checked for
Panther Point, Lewisburg C620, Apollo Lake and Elkhart Lake) the dual
output is enabled when this bit is set and disabled if not. Currently the
logic displays it the other way around when parsing the descriptor.

This patch changes this so now if bit 30 in FLCOMP is not set, dual read
support for fast read operation is shown as disabled.

Original-Change-Id: If6282ac8326ab0b92e9c70c09dba0299bf0deb6f
Original-Signed-off-by: Werner Zeh <werner.zeh@siemens.com>
Original-Reviewed-on: https://review.coreboot.org/c/flashrom/+/61362
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
(cherry picked from commit ff4c624d94d77543313b1ff8247c6430bae14bcd)

BUG=none
BRANCH=none
TEST=make && ninja && ninja test

Signed-off-by: Nikolai Artemiev <nartemiev@google.com>
Change-Id: I9008f14973106b6b98940fb196f54b0527ca5a18
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/3473142
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Tested-by: Edward O'Callaghan <quasisec@chromium.org>