UPSTREAM: libflashrom/fmap: Don't use off_t for flash offsets
off_t is a special POSIX type that is used to represent file offsets in
certain APIs (e.g. lseek(), mmap()), and should not be reused to
represent anything else (such as flash offsets). In particular, the
width of the type may change based on the definition of the
_FILE_OFFSET_BITS macro. Using such a type at the libflashrom interface
is particularly dangerous, because if a program is built with a
different _FILE_OFFSET_BITS value than libflashrom, the resulting ABI
corruption will cause very very nasty and confusing bugs. This patch
replaces all instances of off_t that are not related to file offsets
with (s)size_t.

BUG=b:219811851
TEST=`elogtool list` on cherry.

Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: I68a386973f79ea634f63dfcd7d95a63400e1fdee
Original-Reviewed-on: https://review.coreboot.org/c/flashrom/+/61943
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Original-Reviewed-by: Nikolai Artemiev <nartemiev@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
(cherry picked from commit 016b5abc837cc8ceaa939d1e881cb07a3185cbff)

BRANCH=none
TEST=make && ninja && ninja test

Signed-off-by: Nikolai Artemiev <nartemiev@google.com>
Change-Id: I250232be7bbc13ba57fa8d8a0bd6c3d36e7ba226
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/3476777
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Tested-by: Edward O'Callaghan <quasisec@chromium.org>