CHROMIUM: flashrom: Call cros_ec_finish() after verification
Software write protect on dartmonkey FPMCU board can't be disabled
(without clearing flash content). In this case FPMCU will try to disable
software write protect when hardware write protect is disabled, but it
will always fail to do so, it goes into the bootloop.

To avoid bootloop, a 'power-on' reset flag needs to be present. There is
a workaround in RW which sets the reset flag, but the problem still
occurs when performing reboot from RO. In fact flashrom reboots FPMCU
after flashing RW part (so FPMCU is in RO), but before verification.
This leads to failure in flash verification.

This patch makes cros_ec_finish() to be called after flash was
verified.

This change affects boards with RWSIG task enabled (on ToT or firmware
branch), however there are boards that are not flashed with this tool:
- nucleo-f412zg and nucleo-h743zi a development boards from STM.
- hammer, a detachable keyboard (nocturne), it is updated by hammerd via
  USB which uses it's own implementation.
- prism, a keyboard for vell which is updated using usb_updater2.
- gingerbread, baklava, quiche, panqueque, a honeybuns (docking
  stations) updated using fwupd via USB.

Affected boards that are flashed with this tool:
- nocturne_fp (also known as nami_fp and dartmonkey).
- hatch_fp (also known as bloonchipper).
- rose - eve touchpad firmware.

BUG=b:119131962
BRANCH=none
TEST=Test on Nami board. Compile FPMCU firmware for nami_fp and copy it
     to DUT. Enable hardware write protect using
     'dut-control fw_wp_state:force_on'. Enable software write protect
     using 'ectool --name=cros_fp flashprotect enable' and reboot FPMCU
     using 'ectool --name=cros_fp reboot_ec'. Disable hardware write
     protect using 'dut-control fw_wp_state:force_off'. Run flashrom to
     update FPMCU `flashrom --noverify-all --programmer=ec:type=fp \
     --image=EC_RW  --write <path_to_compiled_firmware>`.

Signed-off-by: Patryk Duda <pdk@semihalf.com>
Change-Id: I530ce0681fcdc7118ce6a0a5b88d1fb31885627f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/3516333
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Tested-by: Patryk Duda <patrykd@google.com>
Reviewed-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Patryk Duda <patrykd@google.com>