dep_writeprotect: drop support for MX25U6435E
It was only ever qualified for BayTrail which is at AUE now. Delete
support to make it easier to achieve feature parity in the new WP code.

The MX25U6435E also doesn't support the ranges that are required for it
to pass flashrom_tester, so it is very unlikley to have been qualified
for anything recent.

The Cq-Depend line below is to ensure that the FlashromWriteProtect
tast test runs against this change.

BUG=b:195381327,b:153800563
BRANCH=none
TEST=CQ

Cq-Depend: chromium:3556867
Signed-off-by: Nikolai Artemiev <nartemiev@google.com>
Change-Id: I272a61a26207d4c8ec1b88b4e90f4cecb40b62a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/3565751
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>