UPSTREAM: flashrom.c, flashcips.c: Test the order of erase functions
Add a check so that the erase functions for all flashchips are in
increasing order of their respective eraseblock sizes. This is required
for the implentation of the improved erasing algorithm. The patch uses
the count of eraseblocks in each erase function to determine the order
(More eraseblocks means that the function has smaller eraseblock size).
Also fix the structs in flashchips.c which were found to be not
conforming to this test.

BUG=none
TEST= make && ./flashrom
also running on octopus dut
flashrom -p host

(cherry picked from commit 7db2baa77d41c3a74449a3f2b907025f69b776b9)

Original-Change-Id: I137cb40483fa690ecc6c7eaece2d9d3f7a851bb4
Original-Signed-off-by: Aarya Chaumal <aarya.chaumal@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/flashrom/+/64961
Original-Reviewed-by: Thomas Heijligen <src@posteo.de>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: 7db2baa77d41c3a74449a3f2b907025f69b776b9
Change-Id: I57c934fbd99b5d11c12d0a128c2213a3316af15e
Signed-off-by: CopyBot <copybot.service@gmail.com>
Signed-off-by: Anastasia Klimchuk <aklm@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/3715978
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>