ichspi,opaque_statusreg: add unlock function for opaque masters
Previously ichspi would use `spi_disable_blockprotect` as the opaque
chip's unlock function, but this was removed in `commit
08fd1b9f557455490345ac42c1c2abe5203ca30f` because it relied on dispatch
hacks in spi25_statusreg.c.

However not having an unlock function causes flashrom to fail
write/erase ops if the flash has an active writeprotect range, e.g.
  flashrom --wp-range 0x0,0x100000
  flashrom -E # <- Fails

This commit creates a copy of `spi_disable_blockprotect` in
opaque_statusreg.c and uses it as the opaque chip's unlock funciton with
ichspi. The new function calls the opaque register access functions
directly, avoiding the need for dispatch hacks in spi25_statusreg.c.

This is still a temporary fix and should be replaced by WP-based
unlocking.

BUG=b:236660711
BRANCH=none
TEST=flashrom --wp-range <non-empty>; flashrom -E on dedede
TEST=firmware.FlashromTester on dedede

Signed-off-by: Nikolai Artemiev <nartemiev@google.com>
Change-Id: Ifd40f245df1af2b6840ade2d02a5d003cd0d748d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/3715730
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>