Makefile: replace with upstream
We do not use the Makefile for Chromium builds, and the divergence
between our tree and upstream causes conflicts while integrating
upstream changes. Since we don't use the Makefile, assume our changes to
it are no longer relevant and simply replace it with upstream's. This
version matches upstream revision
5bb1698e6aa2b41627452ca6e99c878361335c2d.

BUG=b:241308039
TEST=unnecessary
BRANCH=none

Change-Id: I5ddec715df34882803b98a3babf3f2115b88f913
Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/3807023
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Anastasia Klimchuk <aklm@chromium.org>