CHROMIUM: flashchips.c: Fake flash part size as 32MB for Rex Proto 0
Rex Proto 0 has 64MB flash part but the size of image file is 32MB.
Flashrom tool fails to program an image on size mismatch.
Increasing the size of image file will increase the time to program
the flash part which involves the steps like read/write/verify so
this patch tries fake the flash part size as 32MB.

This is a temporary change for Rex Proto 0 and shall be reverted for
Proto 1 which will have 32MB flash part.

BUG=b:242958619,b:243615601
TEST=Verified similar changes on MTL RVP.

Signed-off-by: Kapil Porwal <kapilporwal@google.com>
Change-Id: If187edbe8c7aa3d985734538ec47b2cc494a2336
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/3851163
Reviewed-by: Subrata Banik <subratabanik@chromium.org>
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Nikolai Artemiev <nartemiev@google.com>
Commit-Queue: Kapil Porwal <kapilporwal@chromium.org>